ubuntu 16 for hub base, some scripting

+++ modules/base/configure.sh
@@ -4,6 +4,10 @@ case $VERB in
   "build")
     echo "Building base image kooplex-base"
 
+  if [ -e $DISKIMG/kooplexfs.img ]; then
+   echo "WARNING $DISKIMG/kooplexfs.img exists!!"
+   echo "If you want to remove it, please, use remove and purge for complete removal"
+  else
     mkdir -p $DISKIMG
     cnt=`echo $DISKSIZE_GB | awk '{print $1*1000}'`
     dd if=/dev/zero of=$DISKIMG/kooplexfs.img bs=1M count=$cnt
@@ -15,6 +19,7 @@ case $VERB in
     quotacheck -cuvg $SRV
     quotaon -vu $SRV
     quotaon -vg $SRV
+  fi
     
     mkdir -p $SECRETS
     docker $DOCKERARGS build -t kooplex-base  .
@@ -33,14 +38,15 @@ case $VERB in
     
   ;;
   "remove")
-  umount $DISKIMG/kooplexfs.img 
+
   ;;
   "purge")
+  umount $SRV 
+  echo "Check if $SRV is still mounted! Then run: ' rm -f $DISKIMG/kooplexfs.img '" 
+  #rm -f $DISKIMG/kooplexfs.img 
 
   ;;
   "clean")
-    umount $SRV
-    rm -f $DISKIMG/kooplexfs.img 
     echo "Cleaning base image kooplex-base"
     docker $DOCKERARGS rmi kooplex-base
   ;;
